Visual update: dark background with icon, select options background.
Since the original intent of this app is to drive some mirrors
reflecting lasers, this app is expected to be used in a dark
environment so the color theme is generally dark including background.

While we're adjusting background, the app icon was added to the top
of the background as well.

Waveform select options were also styled so they have dark background
as well. This resolves the issue where options were previously white on
white rendering them unreadable.